requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
nltk>=3.6.0
pydantic>=1.9.0
python-dotenv>=0.19.0
tenacity>=8.0.1 